moving agg_config and agg_configs to ui/agg_types
