+++ src/legacy/ui/public/agg_types/agg_type.ts
@@ -24,7 +24,7 @@ import { AggParam, initParams } from './agg_params';
 // @ts-ignore
 import { FieldFormat, fieldFormats } from '../registry/field_formats';
 import { AggConfig } from '../vis';
-import { AggConfigs } from '../vis/agg_configs';
+import { AggConfigs } from './agg_configs';
 import { SearchSource } from '../courier';
 import { Adapters } from '../inspector';
