+++ src/legacy/ui/public/agg_types/param_types/agg.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { AggConfig } from '../../vis/agg_config';
+import { AggConfig } from '../agg_config';
 import { BaseParamType } from './base';
 
 export class AggParamType extends BaseParamType {
