+++ src/legacy/ui/public/agg_types/param_types/base.ts
@@ -18,7 +18,7 @@
  */
 
 import { AggParam } from '../';
-import { AggConfigs } from '../../vis/agg_configs';
+import { AggConfigs } from '../agg_configs';
 import { AggConfig } from '../../vis';
 import { SearchSource } from '../../courier';
