+++ src/legacy/ui/public/vis/__tests__/_agg_configs.js
@@ -21,9 +21,9 @@ import _ from 'lodash';
 import sinon from 'sinon';
 import expect from '@kbn/expect';
 import ngMock from 'ng_mock';
-import { AggConfig } from '../agg_config';
+import { AggConfig } from '../../agg_types/agg_config';
 import { VisProvider } from '..';
-import { AggConfigs } from '../agg_configs';
+import { AggConfigs } from '../../agg_types/agg_configs';
 import FixturesStubbedLogstashIndexPatternProvider from 'fixtures/stubbed_logstash_index_pattern';
 import { Schemas } from '../editors/default/schemas';
 import { AggGroupNames } from '../editors/default/agg_groups';
