+++ src/legacy/ui/public/vis/index.d.ts
@@ -17,6 +17,6 @@
  * under the License.
  */
 
-export { AggConfig } from './agg_config';
+export { AggConfig } from '../agg_types/agg_config';
 export { Vis, VisProvider, VisParams, VisState } from './vis';
 export { VisualizationController, VisType } from './vis_types/vis_type';
