+++ src/legacy/ui/public/vis/vis.js
@@ -30,7 +30,7 @@
 import { EventEmitter } from 'events';
 import _ from 'lodash';
 import { VisTypesRegistryProvider } from '../registry/vis_types';
-import { AggConfigs } from './agg_configs';
+import { AggConfigs } from '../agg_types/agg_configs';
 import { PersistedState } from '../persisted_state';
 import { FilterBarQueryFilterProvider } from '../filter_manager/query_filter';
 import { updateVisualizationConfig } from './vis_update';
