+++ x-pack/legacy/plugins/maps/public/layers/joins/inner_join.test.js
@@ -13,7 +13,7 @@ jest.mock('ui/vis/editors/default/schemas', () => {
   };
 });
 jest.mock('../../kibana_services', () => {});
-jest.mock('ui/vis/agg_configs', () => {});
+jest.mock('ui/agg_types', () => {});
 jest.mock('ui/timefilter', () => {});
 jest.mock('../vector_layer', () => {});
