+++ x-pack/legacy/plugins/maps/public/layers/sources/es_source.js
@@ -13,7 +13,7 @@ import {
 import { createExtentFilter } from '../../elasticsearch_geo_utils';
 import { timefilter } from 'ui/timefilter';
 import _ from 'lodash';
-import { AggConfigs } from 'ui/vis/agg_configs';
+import { AggConfigs } from 'ui/agg_types';
 import { i18n } from '@kbn/i18n';
 import { ESAggMetricTooltipProperty } from '../tooltips/es_aggmetric_tooltip_property';
