+++ x-pack/legacy/plugins/maps/public/layers/sources/es_term_source.js
@@ -8,7 +8,7 @@ import _ from 'lodash';
 
 import { AbstractESSource } from './es_source';
 import { Schemas } from 'ui/vis/editors/default/schemas';
-import { AggConfigs } from 'ui/vis/agg_configs';
+import { AggConfigs } from 'ui/agg_types';
 import { i18n } from '@kbn/i18n';
 import { ESTooltipProperty } from '../tooltips/es_tooltip_property';
 import { ES_SIZE_LIMIT } from '../../../common/constants';
