+++ x-pack/legacy/plugins/maps/public/layers/sources/es_term_source.test.js
@@ -11,7 +11,7 @@ jest.mock('ui/vis/editors/default/schemas', () => ({
   Schemas: function () {}
 }));
 jest.mock('../../kibana_services', () => {});
-jest.mock('ui/vis/agg_configs', () => {});
+jest.mock('ui/agg_types', () => {});
 jest.mock('ui/timefilter', () => {});
 
 const indexPatternTitle = 'myIndex';
